+++ apps/remnant2toolkit/app/(builds)/_hooks/use-build-list-state.ts
@@ -47,6 +47,7 @@ const DEFAULT_STATE: State = {
     validatedViewCount: 0,
     duplicateCount: 0,
     buildItems: [],
+    percentageOwned: 0,
   })),
   totalBuildCount: 0,
   isLoading: false,