+++ apps/remnant2toolkit/app/(builds)/_types/db-build.ts
@@ -32,4 +32,5 @@ export interface DBBuild {
   variantIndex: number;
   videoUrl: Build['videoUrl'];
   viewCount: number;
+  percentageOwned: number;
 }