+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-09-30
+
+### Added
+
+- Added the ability to sort builds by percentage of items owned. Update your items in the Item Tracker, and then use the new build filter!
+- Added more new items.
+
 ## 2024-09-29
 
 ### Added